Adding Hybrid Radio Core
This patch brings in the addition of HybridRadioCore
class. The intention is to be able to use this as plug
and play to any module and not be dependent if the
application is UI based or command line based.
Eventually we will be moving this to a library.

This patch also brings in some clean up of the code
and adds DAB addition for DAB station Antenna
Bayern from Germany.